Use use_table_data
